ENH: Add data fetcher
Add data fetcher. Add the accompanying test.

Add a script to fetch a given dataset to the provided local path. Add
the accompanying test. Add the script to the project scripts in the
`pyproject.toml` file.

Add the badge corresponding to the Zenodo record DOI to the `README`.

Explain in the `README` file how to:
- Fetch the data automatically.
- Cite the data.

Enable the package testing and coverage report steps in the
`GitHub Actions` workflow file.